+++ RraptorPult/src/com/rraptor/pult/model/Point2D.java
@@ -4,45 +4,45 @@ import java.io.Serializable;
 
 public class Point2D implements Serializable {
 
-	private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-	private int x;
-	private int y;
+    private double x;
+    private double y;
 
-	public Point2D(int x, int y) {
-		super();
-		this.x = x;
-		this.y = y;
-	}
+    public Point2D(double x, double y) {
+        super();
+        this.x = x;
+        this.y = y;
+    }
 
-	@Override
-	public boolean equals(Object o) {
-		if (o instanceof Point2D) {
-			final Point2D p2 = (Point2D) o;
-			return (p2.getX() == x) && (p2.getY() == y);
-		} else {
-			return false;
-		}
-	}
+    @Override
+    public boolean equals(Object o) {
+        if (o instanceof Point2D) {
+            final Point2D p2 = (Point2D) o;
+            return (p2.getX() == x) && (p2.getY() == y);
+        } else {
+            return false;
+        }
+    }
 
-	public int getX() {
-		return x;
-	}
+    public double getX() {
+        return x;
+    }
 
-	public int getY() {
-		return y;
-	}
+    public double getY() {
+        return y;
+    }
 
-	@Override
-	public int hashCode() {
-		return x + y;
-	}
+    @Override
+    public int hashCode() {
+        return (int) (x + y);
+    }
 
-	public void setX(int x) {
-		this.x = x;
-	}
+    public void setX(double x) {
+        this.x = x;
+    }
 
-	public void setY(int y) {
-		this.y = y;
-	}
+    public void setY(double y) {
+        this.y = y;
+    }
 }
